+++ config/config.go
@@ -87,10 +87,9 @@ var (
 	errUnmarshalling                          = errors.New("unmarshalling failed")
 	errFileDoesNotExist                       = errors.New("file does not exist")
 	errInvalidSignerConfig                    = fmt.Errorf("only one of the following flags can be set: %s, %s, %s, %s", StakingEphemeralSignerEnabledKey, StakingSignerKeyContentKey, StakingSignerKeyPathKey, StakingRPCSignerEndpointKey)
-	errDiskSpaceOutOfRange                    = fmt.Errorf("out of range [0,%d]", maxDiskSpaceThreshold)
-	errDiskWarnAfterFatal                     = errors.New("warning disk space threshold cannot be greater than fatal threshold")
-	errMemorySpaceOutOfRange                  = fmt.Errorf("out of range [0,%d]", maxMemorySpaceThreshold)
-	errMemoryWarnAfterFatal                   = errors.New("warning memory threshold cannot be greater than fatal threshold")
+	errDiskSpaceOutOfRange   = fmt.Errorf("out of range [0,%d]", maxDiskSpaceThreshold)
+	errDiskWarnAfterFatal    = errors.New("warning disk space threshold cannot be greater than fatal threshold")
+	errMemorySpaceOutOfRange = fmt.Errorf("out of range [0,%d]", maxMemorySpaceThreshold)
 )
 
 func getConsensusConfig(v *viper.Viper) snowball.Parameters {
@@ -1162,26 +1161,14 @@ func getDiskSpaceConfig(v *viper.Viper) (
 	}
 }
 
-func getMemoryConfig(v *viper.Viper) (
-	requiredAvailableMemoryPercentage uint64,
-	warningAvailableMemoryPercentage uint64,
-	err error,
-) {
-	var (
-		warnKey     = SystemTrackerWarningAvailableMemoryPercentageKey
-		requiredKey = SystemTrackerRequiredAvailableMemoryPercentageKey
+func getMemoryConfig(v *viper.Viper) (uint64, error) {
+	warnKey := SystemTrackerWarningAvailableMemoryPercentageKey
+	warn := v.GetUint64(warnKey)
 
-		warn     = v.GetUint64(warnKey)
-		required = v.GetUint64(requiredKey)
-	)
-	switch {
-	case warn > maxMemorySpaceThreshold:
-		return 0, 0, fmt.Errorf("%w: %q (%d)", errMemorySpaceOutOfRange, warnKey, warn)
-	case warn < required:
-		return 0, 0, fmt.Errorf("%w: %d < %d", errMemoryWarnAfterFatal, warn, required)
-	default:
-		return required, warn, nil
+	if warn > maxMemorySpaceThreshold {
+		return 0, fmt.Errorf("%w: %q (%d)", errMemorySpaceOutOfRange, warnKey, warn)
 	}
+	return warn, nil
 }
 
 func getDiskTargeterConfig(v *viper.Viper) (tracker.TargeterConfig, error) {
@@ -1441,7 +1428,7 @@ func GetNodeConfig(v *viper.Viper) (node.Config, error) {
 		return node.Config{}, err
 	}
 
-	nodeConfig.RequiredAvailableMemoryPercentage, nodeConfig.WarningAvailableMemoryPercentage, err = getMemoryConfig(v)
+	nodeConfig.WarningAvailableMemoryPercentage, err = getMemoryConfig(v)
 	if err != nil {
 		return node.Config{}, err
 	}