+++ config/node/config.go
@@ -215,8 +215,7 @@ type Config struct {
 	RequiredAvailableDiskSpacePercentage uint64 `json:"requiredAvailableDiskSpacePercentage"`
 	WarningAvailableDiskSpacePercentage  uint64 `json:"warningAvailableDiskSpacePercentage"`
 
-	RequiredAvailableMemoryPercentage uint64 `json:"requiredAvailableMemoryPercentage"`
-	WarningAvailableMemoryPercentage  uint64 `json:"warningAvailableMemoryPercentage"`
+	WarningAvailableMemoryPercentage uint64 `json:"warningAvailableMemoryPercentage"`
 
 	TraceConfig trace.Config `json:"traceConfig"`